backend_requirements.txt
# Backend Requirements for Google Cloud Video Breakdown System

# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Google Cloud libraries
google-cloud-speech==2.21.0
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.38.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Audio/Video processing
ffmpeg-python==0.2.0

# Utilities
python-dotenv==1.0.0
uuid==1.30
requests==2.31.0

# Optional: For better JSON handling
orjson==3.9.10

# Optional: For logging and monitoring
structlog==23.2.0 